refactor(lib): reorganize content collections schema and loader
