+++ src/content.config.ts
@@ -1,14 +1,16 @@
-import { authors } from "./lib/astro/collections/schema/authors";
-import { blogCategories } from "./lib/astro/collections/schema/blog-categories";
-import { blogPosts } from "./lib/astro/collections/schema/blog-posts";
-import { blogroll } from "./lib/astro/collections/schema/blogroll";
-import { bookmarks } from "./lib/astro/collections/schema/bookmarks";
-import { guides } from "./lib/astro/collections/schema/guides";
-import { indexPages } from "./lib/astro/collections/schema/index-pages";
-import { notes } from "./lib/astro/collections/schema/notes";
-import { pages } from "./lib/astro/collections/schema/pages";
-import { projects } from "./lib/astro/collections/schema/projects";
-import { tags } from "./lib/astro/collections/schema/tags";
+import {
+  authors,
+  blogCategories,
+  blogPosts,
+  blogroll,
+  bookmarks,
+  guides,
+  indexPages,
+  notes,
+  pages,
+  projects,
+  tags,
+} from "./lib/astro/schema";
 
 export const collections = {
   authors,
