--- src/lib/astro/collections/schema/tags.ts
@@ -1,27 +1,0 @@
-import { defineCollection, reference, z } from "astro:content";
-import { isAvailableLocale } from "../../../../utils/type-guards";
-import { globLoader } from "../../loaders/glob-loader";
-import { contentsBaseSchema, coverSchema } from "./partials";
-
-export const tags = defineCollection({
-  loader: globLoader("tags"),
-  schema: ({ image }) =>
-    contentsBaseSchema
-      .extend({
-        cover: coverSchema(image).optional(),
-        i18n: z
-          .record(z.string().refine(isAvailableLocale), reference("tags"))
-          .optional(),
-        permaslug: z.string().optional(),
-      })
-      .transform(({ isDraft, publishedOn, updatedOn, ...tag }) => {
-        return {
-          ...tag,
-          meta: {
-            isDraft,
-            publishedOn,
-            updatedOn: updatedOn ?? publishedOn,
-          },
-        };
-      }),
-});
