+++ src/lib/astro/loaders/index.ts
@@ -1,0 +1,1 @@
+export * from "./glob-loader";
