+++ src/lib/astro/schema/blog-categories.test.ts
@@ -1,11 +1,10 @@
 import { describe, expect, it, vi } from "vitest";
-import { createReferenceMock } from "../../../../../tests/mocks/references";
-import { createImageMock } from "../../../../../tests/mocks/schema";
-import { CONFIG } from "../../../../utils/constants";
-import { guides } from "./guides";
+import { createImageMock } from "../../../../tests/mocks/schema";
+import { CONFIG } from "../../../utils/constants";
+import { blogCategories } from "./blog-categories";
 
-vi.mock("../../../../utils/dates", async (importOriginal) => {
-  const mod = await importOriginal<typeof import("../../../../utils/dates")>();
+vi.mock("../../../utils/dates", async (importOriginal) => {
+  const mod = await importOriginal<typeof import("../../../utils/dates")>();
 
   return {
     ...mod,
@@ -13,27 +12,18 @@ vi.mock("../../../../utils/dates", async (importOriginal) => {
   };
 });
 
-vi.mock("astro:content", async () => {
-  const actual = await vi.importActual("astro:content");
-  return {
-    ...actual,
-    reference: vi.fn((collection: string) => createReferenceMock(collection)),
-  };
-});
-
 const mockImage = createImageMock();
 
-describe("guides", () => {
+describe("blogCategories", () => {
   it("should include the meta in the transformed output", async () => {
     /* eslint-disable-next-line @typescript-eslint/no-magic-numbers -- Self-explanatory. */
     expect.assertions(4);
 
-    const guide = {
-      title: "The title of the guide",
-      description: "A description of the guide.",
+    const category = {
+      title: "The title of the category",
+      description: "A description of the category.",
       isDraft: true,
       publishedOn: new Date("2023-01-01"),
-      authors: ["john-doe"],
       seo: {
         title: "qui sit vero",
         description: "Vel voluptatem laboriosam.",
@@ -41,12 +31,12 @@ describe("guides", () => {
       updatedOn: new Date("2023-01-02"),
     };
 
-    if (typeof guides.schema !== "function") {
+    if (typeof blogCategories.schema !== "function") {
       throw new TypeError("The schema is not callable");
     }
 
-    const parsedSchema = guides.schema({ image: mockImage });
-    const result = await parsedSchema.safeParseAsync(guide);
+    const parsedSchema = blogCategories.schema({ image: mockImage });
+    const result = await parsedSchema.safeParseAsync(category);
 
     expect(result.success).toBe(true);
 
@@ -62,23 +52,22 @@ describe("guides", () => {
     /* eslint-disable-next-line @typescript-eslint/no-magic-numbers -- Self-explanatory. */
     expect.assertions(4);
 
-    const guide = {
-      title: "The title of the guide",
-      description: "A description of the guide.",
+    const category = {
+      title: "The title of the category",
+      description: "A description of the category.",
       publishedOn: new Date("2023-01-01"),
-      authors: ["john-doe"],
       seo: {
         title: "qui sit vero",
         description: "Vel voluptatem laboriosam.",
       },
     };
 
-    if (typeof guides.schema !== "function") {
+    if (typeof blogCategories.schema !== "function") {
       throw new TypeError("The schema is not callable");
     }
 
-    const parsedSchema = guides.schema({ image: mockImage });
-    const result = await parsedSchema.safeParseAsync(guide);
+    const parsedSchema = blogCategories.schema({ image: mockImage });
+    const result = await parsedSchema.safeParseAsync(category);
 
     expect(result.success).toBe(true);
