+++ src/lib/astro/schema/blog-posts.ts
@@ -1,20 +1,15 @@
 import { defineCollection, reference, z } from "astro:content";
-import { isAvailableLocale } from "../../../../utils/type-guards";
-import { globLoader } from "../../loaders/glob-loader";
-import { contentsBaseSchema, coverSchema } from "./partials";
+import { globLoader } from "../loaders";
+import { contentsBaseSchema, i18nSchema } from "./partials";
 
 export const blogPosts = defineCollection({
   loader: globLoader("blog.posts"),
   schema: ({ image }) =>
-    contentsBaseSchema
+    contentsBaseSchema(image)
       .extend({
         authors: z.array(reference("authors")),
         category: reference("blog.categories"),
-        cover: coverSchema(image).optional(),
-        i18n: z
-          .record(z.string().refine(isAvailableLocale), reference("blog.posts"))
-          .optional(),
-        permaslug: z.string().optional(),
+        i18n: i18nSchema("blog.posts").optional(),
         tags: z.array(reference("tags")).optional(),
       })
       .transform(
