+++ src/lib/astro/schema/index.ts
@@ -1,0 +1,11 @@
+export * from "./authors";
+export * from "./blog-categories";
+export * from "./blog-posts";
+export * from "./blogroll";
+export * from "./bookmarks";
+export * from "./guides";
+export * from "./index-pages";
+export * from "./notes";
+export * from "./pages";
+export * from "./projects";
+export * from "./tags";
