+++ src/lib/astro/schema/tags.test.ts
@@ -1,10 +1,10 @@
 import { describe, expect, it, vi } from "vitest";
-import { createImageMock } from "../../../../../tests/mocks/schema";
-import { CONFIG } from "../../../../utils/constants";
-import { blogCategories } from "./blog-categories";
+import { createImageMock } from "../../../../tests/mocks/schema";
+import { CONFIG } from "../../../utils/constants";
+import { tags } from "./tags";
 
-vi.mock("../../../../utils/dates", async (importOriginal) => {
-  const mod = await importOriginal<typeof import("../../../../utils/dates")>();
+vi.mock("../../../utils/dates", async (importOriginal) => {
+  const mod = await importOriginal<typeof import("../../../utils/dates")>();
 
   return {
     ...mod,
@@ -14,14 +14,14 @@ vi.mock("../../../../utils/dates", async (importOriginal) => {
 
 const mockImage = createImageMock();
 
-describe("blogCategories", () => {
+describe("tags", () => {
   it("should include the meta in the transformed output", async () => {
     /* eslint-disable-next-line @typescript-eslint/no-magic-numbers -- Self-explanatory. */
     expect.assertions(4);
 
-    const category = {
-      title: "The title of the category",
-      description: "A description of the category.",
+    const tag = {
+      title: "The title of the tag",
+      description: "A description of the tag.",
       isDraft: true,
       publishedOn: new Date("2023-01-01"),
       seo: {
@@ -31,12 +31,12 @@ describe("blogCategories", () => {
       updatedOn: new Date("2023-01-02"),
     };
 
-    if (typeof blogCategories.schema !== "function") {
+    if (typeof tags.schema !== "function") {
       throw new TypeError("The schema is not callable");
     }
 
-    const parsedSchema = blogCategories.schema({ image: mockImage });
-    const result = await parsedSchema.safeParseAsync(category);
+    const parsedSchema = tags.schema({ image: mockImage });
+    const result = await parsedSchema.safeParseAsync(tag);
 
     expect(result.success).toBe(true);
 
@@ -52,9 +52,9 @@ describe("blogCategories", () => {
     /* eslint-disable-next-line @typescript-eslint/no-magic-numbers -- Self-explanatory. */
     expect.assertions(4);
 
-    const category = {
-      title: "The title of the category",
-      description: "A description of the category.",
+    const tag = {
+      title: "The title of the tag",
+      description: "A description of the tag.",
       publishedOn: new Date("2023-01-01"),
       seo: {
         title: "qui sit vero",
@@ -62,12 +62,12 @@ describe("blogCategories", () => {
       },
     };
 
-    if (typeof blogCategories.schema !== "function") {
+    if (typeof tags.schema !== "function") {
       throw new TypeError("The schema is not callable");
     }
 
-    const parsedSchema = blogCategories.schema({ image: mockImage });
-    const result = await parsedSchema.safeParseAsync(category);
+    const parsedSchema = tags.schema({ image: mockImage });
+    const result = await parsedSchema.safeParseAsync(tag);
 
     expect(result.success).toBe(true);
